move read_image to utils
